+++ seafoam/__init__.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*- #
 """
-MinchinDotCa
-=============
+Seafoam
+=======
 
 This is a Pelican theme. This is the helper code to go with it.
 """
 
 from pathlib import Path
 
-__version__ = "1.2.0+dev.0"
+__version__ = "2.0.0+dev.0"
 
 
 def get_path():